[refactor]: split consultation app service by workflow responsibility
